- revert "final" due to limited compiler support and make destructors virtual for safety - move rapidxml back into main src dir due to CRAN requirements
