+++ src/rapidxml.h
